agent/requirements.txt
# FastAPI + server
fastapi==0.136.0
uvicorn[standard]==0.44.0
python-multipart==0.0.20

# Pydantic v2 (required by FastAPI and LangChain 1.0)
pydantic==2.13.2
pydantic-settings==2.9.1

# HTTP client — httpx only, never requests (see CLAUDE.md §Forbidden)
httpx==0.28.1

# LangChain / LangGraph
langchain==1.2.15
langchain-core==1.3.0
langchain-anthropic==1.4.1
langgraph==1.1.8
langgraph-prebuilt==1.0.10
langgraph-checkpoint==4.0.2
langgraph-checkpoint-postgres==3.0.5

# Postgres driver — psycopg3 async (required by langgraph-checkpoint-postgres)
psycopg[binary,pool]==3.3.3

# pgvector support for PostgresStore
pgvector==0.4.2

# JWT verification — PyJWT (simpler than python-jose)
PyJWT==2.12.1

# Async utilities
anyio==4.13.0

# Test dependencies (respx requires httpx, already pinned above)
pytest==9.0.3
pytest-asyncio==1.3.0
respx==0.23.1
fakeredis==2.27.0  # M5: in-memory Redis for BudgetGuard tests

## M5: routing & budget guard ------------------------------------------------
# Multi-provider chat models for the routing chain (Anthropic primary; Gemini
# and OpenRouter fallbacks on infra failure only — see model_router.py).
langchain-google-genai==2.1.5
langchain-openai==0.4.4
# Async Redis client used by BudgetGuardMiddleware to read Spring's usage counters.
# Same key shape as RedisUsageCounters.java (ai:usage:{userId}:day:{YYYY-MM-DD}:usd).
redis==5.4.0

## M5: observability --------------------------------------------------------
# OpenTelemetry SDK + OTLP gRPC exporter for traces/metrics → otel-collector
# → Langfuse (traces) + Prometheus (metrics). Auto-instrumentation for FastAPI
# and httpx so spans cover the inbound SSE endpoint and outbound tool calls.
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-httpx==0.49b2
# Langfuse Python SDK (LangChain-native CallbackHandler) for richer trace UI.
langfuse==3.4.0
